requirements.txt
coverage==7.6.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
